requirements.txt
torch
torchvision
numpy
scipy
Pillow
scikit-learn
tqdm